# application/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Mechanic API
  description: |
    A comprehensive REST API for managing auto repair shop operations including mechanics, customers, inventory, and service tickets.
    
    ## Authentication
    This API uses JWT (JSON Web Tokens) for authentication. After logging in, include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## Rate Limiting
    - Customer endpoints: 100 requests per hour
    - Login endpoints: 10 requests per minute
    - Specific endpoints may have additional rate limits
    
    ## Error Handling
    The API uses standard HTTP status codes and returns error messages in JSON format.
  version: 1.0.0
  contact:
    name: API Support
    email: support@mechanicapi.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.mechanic.com/v1
    description: Production server
...
paths:
...
  /service-tickets/:
    get:
      tags:
        - Service Tickets
      summary: Get all service tickets
      description: Retrieve a list of all service tickets
      responses:
        '200':
          description: Service tickets retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServiceTicket'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Service Tickets
      summary: Create a new service ticket
      description: Create a new service ticket (requires customer authentication)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceTicketCreate'
      responses:
        '201':
          description: Service ticket created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceTicket'
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /service-tickets/{ticket_id}:
    get:
      tags:
        - Service Tickets
      summary: Get specific service ticket by ID
      description: Retrieve a specific service ticket by its ID
      parameters:
        - name: ticket_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Service ticket retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceTicket'
        '404':
          description: Ticket not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Service Tickets
      summary: Update service ticket
      description: Update a specific service ticket by its ID (requires mechanic authentication)
      security:
        - BearerAuth: []
      parameters:
        - name: ticket_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceTicketUpdate'
      responses:
        '200':
          description: Service ticket updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceTicket'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Ticket not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Service Tickets
      summary: Delete service ticket
      description: Delete a specific service ticket by its ID (requires mechanic authentication)
      security:
        - BearerAuth: []
      parameters:
        - name: ticket_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '204':
          description: Service ticket deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Ticket not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /service-tickets/{ticket_id}/edit:
    put:
      tags:
        - Service Tickets
      summary: Update mechanics assigned to a service ticket
      description: Add or remove mechanics from a service ticket (requires mechanic authentication)
      security:
        - BearerAuth: []
      parameters:
        - name: ticket_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                remove_ids:
                  type: array
                  items:
                    type: integer
                  example: [2, 3]
                  description: List of mechanic IDs to remove from ticket
                add_ids:
                  type: array
                  items:
                    type: integer
                  example: [4, 5]
                  description: List of mechanic IDs to add to ticket
      responses:
        '200':
          description: Ticket mechanics updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceTicket'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Ticket not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /service-tickets/{ticket_id}/parts:
    post:
      tags:
        - Service Tickets
      summary: Add parts/inventory to service ticket
      description: Add parts or inventory items to a service ticket (requires mechanic authentication)
      security:
        - BearerAuth: []
      parameters:
        - name: ticket_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - part_id
              properties:
                part_id:
                  type: integer
                  example: 1
                  description: Inventory item ID to add to ticket
                quantity:
                  type: integer
                  example: 2
                  description: Quantity of the part to use
                  default: 1
      responses:
        '200':
          description: Part added to ticket successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceTicket'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Ticket or part not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Health Check
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoints
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Registration failed"
        message:
          type: string
          example: "Internal server error occurred"
        details:
          type: string
          example: "Database connection error"
...
    MechanicResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        specialization:
          type: string
          example: "Engine Repair"
        phone:
          type: string
          example: "123-456-7890"
        address:
          type: string
          example: "123 Main St, City, State"
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
...
    ServiceTicket:
      type: object
      properties:
        id:
          type: integer
          example: 1
        customer_id:
          type: integer
          example: 1
        description:
          type: string
          example: "Engine overheating and strange noises"
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
          example: "pending"
        priority:
          type: string
          enum: [low, normal, high, urgent]
          example: "normal"
        vehicle_info:
          type: string
          example: "2020 Toyota Camry, VIN: 123456789"
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T14:45:00Z"
        mechanics:
          type: array
          items:
            $ref: '#/components/schemas/MechanicResponse'
        parts:
          type: array
          items:
            type: object
            properties:
              part_id:
                type: integer
                example: 1
              name:
                type: string
                example: "Engine Oil"
              quantity:
                type: integer
                example: 2
    ServiceTicketCreate:
      type: object
      required:
        - description
      properties:
        description:
          type: string
          example: "Engine overheating and strange noises"
        priority:
          type: string
          enum: [low, normal, high, urgent]
          example: "normal"
          default: "normal"
        vehicle_info:
          type: string
          example: "2020 Toyota Camry, VIN: 123456789"
    ServiceTicketUpdate:
      type: object
      properties:
        description:
          type: string
          example: "Updated description of the issue"
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
          example: "in_progress"
        priority:
          type: string
          enum: [low, normal, high, urgent]
          example: "high"